np-47035: external sources conversion in own class
